+++ src/mem/mem_set.rs
@@ -7,6 +7,11 @@ pub fn _memset_impl(buf: &mut [u8], c: u8, n: usize) -> Result<(), RangeError> {
     }
     for i in 0..n {
         buf[i] = c;
+        //
+        // should not use unsafe for auto-vector
+        //let item = unsafe { buf.get_unchecked_mut(i) };
+        //*item = c;
+        //
     }
     Ok(())
 }