+++ CHANGELOG.md
@@ -2,6 +2,9 @@ TBD
 ===
 Unreleased changes. Release notes have not yet been written.
 
+* refactering source code
+* add memcmp(), memeq(), memset()
+
 0.1.0 (2021-05-12)
 =====